backend/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
black==24.1.1
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
flake8==7.0.0
greenlet==3.0.1
h11==0.14.0
idna==3.4
isort==5.13.2
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.1.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.6
python-slugify==8.0.4
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
text-unidecode==1.3
typing_extensions==4.8.0
uvicorn==0.24.0.post1
